% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wiod.R
\name{fit_wiod}
\alias{fit_wiod}
\title{WIOD Bayesian modeling}
\usage{
fit_wiod(statistic, levels_id_vector = 0, model, pooling)
}
\arguments{
\item{statistic}{this is the data}

\item{levels_id_vector}{is the vector that shows which subsample a particular observation belongs to}

\item{model}{is the type of model you want to run, choose from normal, lognormal, gamma, Weibull and exponential}

\item{pooling}{is the level of pooling desired, choose from complete, partial or none}
}
\description{
This function allows you to run a variety of Bayesian (hierarchical) models on WIOD data and key statistics to unearth persistent statistical patterns across time and countries
}
\examples{
sample examples here

}
\keyword{wiodbayes}
