% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_wiod.R
\name{loo_wiod}
\alias{loo_wiod}
\title{Leave one out Cross Validation for WIOD Bayesian modeling}
\usage{
loo_wiod(stanfit_object)
}
\arguments{
\item{stanfit_object}{this is the stanfit object that is the argument for the loo-cv function}
}
\description{
This function allows you to compare Bayesian models using leave one out cross validation
}
\examples{
sample examples here

}
